+++ woom/workflow.py
@@ -561,6 +561,7 @@ class Workflow:
 
         content += "-" * 50
 
+        self.logger.info(f"Fake submission: {self.context['task_path']} → {cmdline}")
         self.logger.debug(content)
 
         # Create fake job